Port to Chia 2.7.3 base
Reproducible pipeline: /media/sgroiwes/nvme/port/{rename.py,remerge.sh}
- Retire hand-built cactus_rs fork; use upstream network-agnostic chia_rs 0.45.x
- Consensus identity preserved in cactus/consensus/default_constants.py + initial-config.yaml
- Fix invalid 65-char testnet prefarm hashes (old sed artifact)
- Fix stale chia-named systemd assets (old sed artifact)
- DAO wallet removed (upstream removal in 2.6)

+++ cactus/consensus/block_rewards.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from cactus.util.ints import uint32, uint64
+from chia_rs.sized_ints import uint32, uint64
 
 # 1 Cactus coin = 1,000,000,000,000 = 1 trillion mojo.
 _mojo_per_cactus = 1000000000000
@@ -17,17 +17,17 @@ def calculate_pool_reward(height: uint32) -> uint64:
     """
 
     if height == 0:
-        return uint64(int((7 / 8) * 210000 * _mojo_per_cactus))
+        return uint64((7 / 8) * 210000 * _mojo_per_cactus)
     elif height < 3 * _blocks_per_year:
-        return uint64(int((7 / 8) * 2 * _mojo_per_cactus))
+        return uint64((7 / 8) * 2 * _mojo_per_cactus)
     elif height < 6 * _blocks_per_year:
-        return uint64(int((7 / 8) * 1 * _mojo_per_cactus))
+        return uint64((7 / 8) * 1 * _mojo_per_cactus)
     elif height < 9 * _blocks_per_year:
-        return uint64(int((7 / 8) * 0.5 * _mojo_per_cactus))
+        return uint64((7 / 8) * 0.5 * _mojo_per_cactus)
     elif height < 12 * _blocks_per_year:
-        return uint64(int((7 / 8) * 0.25 * _mojo_per_cactus))
+        return uint64((7 / 8) * 0.25 * _mojo_per_cactus)
     else:
-        return uint64(int((7 / 8) * 0.125 * _mojo_per_cactus))
+        return uint64((7 / 8) * 0.125 * _mojo_per_cactus)
 
 
 def calculate_base_farmer_reward(height: uint32) -> uint64:
@@ -40,14 +40,14 @@ def calculate_base_farmer_reward(height: uint32) -> uint64:
     rates increase continuously.
     """
     if height == 0:
-        return uint64(int((1 / 8) * 210000 * _mojo_per_cactus))
+        return uint64((1 / 8) * 210000 * _mojo_per_cactus)
     elif height < 3 * _blocks_per_year:
-        return uint64(int((1 / 8) * 2 * _mojo_per_cactus))
+        return uint64((1 / 8) * 2 * _mojo_per_cactus)
     elif height < 6 * _blocks_per_year:
-        return uint64(int((1 / 8) * 1 * _mojo_per_cactus))
+        return uint64((1 / 8) * 1 * _mojo_per_cactus)
     elif height < 9 * _blocks_per_year:
-        return uint64(int((1 / 8) * 0.5 * _mojo_per_cactus))
+        return uint64((1 / 8) * 0.5 * _mojo_per_cactus)
     elif height < 12 * _blocks_per_year:
-        return uint64(int((1 / 8) * 0.25 * _mojo_per_cactus))
+        return uint64((1 / 8) * 0.25 * _mojo_per_cactus)
     else:
-        return uint64(int((1 / 8) * 0.125 * _mojo_per_cactus))
+        return uint64((1 / 8) * 0.125 * _mojo_per_cactus)
